Release/1.6.3 (#72)
* Added example Generic CMakeLists.txt file and corresponding cpp file for example usage

* validated airtree-bench/tools/generate.sh and data.sh. Updated the README.md to reflect.

* New script to download FCBench data. Uses gdown to download from GoogleDrive

* New getFCBenchData.sh. Updated comments in script to reflect correct urls

* Changed name to match format throughout repo getFCBenchData.sh -> get_fcbench_data.sh

* Updated the generate.sh benchmark pipeline with supporting scripts. Updated airtree-bench/../generate/NDxF/T/P files for new benchmark variables.

* AirTree Bench -> AirTree Benchmark

* changed a variable that was hiding a class member (serializedTrie)

* Updated generate.sh pipeline to get a bigger dataset for benchmarking

* removed old/renamed scripts

* updated generate.sh to use yellow_tripdata dataset and fixed bug in get_bench_datasets.sh

* get_bench_data.sh replaces data.sh

* fixed small bug in generate.sh and updated output formats

* updated benchmark workflow with query IDs and return value size in output (#69)

* Feature/query improvement (#71)

* add more query bench

* change loglevel to trace

* optimize recostruct and table generation

* add sparse tree walk

* optimize deserialization

* add support for n-way merge

* hard cap max bitlength

* minor addition

* ci fixes

* pre-fetch the key

* windows fix

* prepare release 1.6.3

---------

+++ airtree-bench/src/generate/trie2d/2DxP.cpp
@@ -24,17 +24,17 @@ BENCHMARK_F(AirTreeBench2DxP, Serialize_float)(benchmark::State &state) {
 }
 
 BENCHMARK_F(AirTreeBench2DxP, CreateAndInsert_int32)(benchmark::State &state) {
-  runCreateAndInsert<float>(state);
+  runCreateAndInsert<int32_t>(state);
 }
 
 BENCHMARK_F(AirTreeBench2DxP, Serialize_int32)(benchmark::State &state) {
-  runSerialize<float>(state);
+  runSerialize<int32_t>(state);
 }
 
 BENCHMARK_F(AirTreeBench2DxP, CreateAndInsert_int64)(benchmark::State &state) {
-  runCreateAndInsert<float>(state);
+  runCreateAndInsert<int64_t>(state);
 }
 
 BENCHMARK_F(AirTreeBench2DxP, Serialize_int64)(benchmark::State &state) {
-  runSerialize<float>(state);
+  runSerialize<int64_t>(state);
 }
